pyproject.toml
[tool.poetry]
name = "tvm-news"
version = "0.1.0"
description = ""
authors = ["Jared Roesch <jroesch@octoml.ai>"]

[tool.poetry.dependencies]
python = "^3.7"
termcolor = "^1.1"
PyGithub = "^1.45"
pystache = "^0.5.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
tvm-news = 'tvm_news:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

